chore : adjust url (userlist X, members O)
