+++ src/app/(etc)/members/[id]/page.tsx
@@ -1,9 +1,23 @@
-const PerformanceDetail = () => {
-  return (
-    <>
-      공연 상세보기
-    </>
-  )
+import React from 'react';
+import DescribeToJSX from '@/components/common/DescribeToJSX';
+import { createUser } from '@/lib/dummy/User';
+
+interface Props {
+  params: {
+    id: number;
+  };
 }
 
-export default PerformanceDetail
+const UserDetail = ({ params }: Props) => {
+  const { id } = params;
+  const user = createUser(id)
+  
+  return (
+    <div>
+      <h2>User ID: {id}</h2>
+      <DescribeToJSX data={user} level={0} />
+    </div>
+  );
+};
+
+export default UserDetail;
