+++ src/components/common/Header.tsx
@@ -9,7 +9,7 @@ export const HeaderInner = () => {
     { name: "공지사항", url: ROUTES.NOTICE.LIST.url, active: false },
     { name: "공연목록", url: ROUTES.PERFORMANCE.LIST.url, active: true },
     { name: "세션지원", url: ROUTES.TEAM.LIST.url, active: true },
-    { name: "부원목록", url: ROUTES.USERLIST.LIST.url, active: true},
+    { name: "부원목록", url: ROUTES.MEMBERS.LIST.url, active: true},
   ]
 
   return (
